Adding some generic layout handling

+++ client/systems/movement.py
@@ -12,11 +12,11 @@ class MenuMovementSystem(sdl2.ext.Applicator):
         for velocity, sprite in components:
             curr_time = time.time()
             time_diff = curr_time - velocity.last_update
-            d_x, d_y = (int(velocity.velocity_coords[0] * time_diff), int(velocity.velocity_coords[1] * time_diff))
+            d_x, d_y = (velocity.velocity_coords[0] * time_diff, velocity.velocity_coords[1] * time_diff)
             o_x, o_y = sprite.position
             n_x, n_y = (
-                o_x + d_x,
-                o_y + d_y
+                int(o_x + d_x),
+                int(o_y + d_y)
             )
 
             if velocity.target_coords:
